+++ src/Components/Nav/Nav.jsx
@@ -1,90 +1,89 @@
-import clsx from "clsx";
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { useTranslation } from "react-i18next";
-import { useSelector } from "react-redux";
 
 import AddIcon from "@material-ui/icons/Add";
 import BarChartIcon from "@material-ui/icons/BarChart";
 import LibraryBooksIcon from "@material-ui/icons/LibraryBooks";
 import LocalLibraryIcon from "@material-ui/icons/LocalLibrary";
-import SettingsIcon from "@material-ui/icons/Settings";
+import PolicyIcon from "@material-ui/icons/Policy";
+import SecurityIcon from "@material-ui/icons/Security";
 import StyleIcon from "@material-ui/icons/Style";
+import VerifiedUserIcon from "@material-ui/icons/VerifiedUser";
 
 import NavButton from "./NavButton.jsx";
 
+import useLinkCreator from "../../hooks/useLinkCreator.js";
+
 import "./Nav.scss";
+
+import { pages } from "../../utils/constants";
 
 const Nav = () => {
   const { t } = useTranslation();
-  const menuStyle = useSelector((state) => state.setting.menuStyle);
-  const [initialStyle, setInitialStyle] = useState(menuStyle);
-  const [loading, setLoading] = useState(false);
 
-  const navLayoutClasses = clsx(
-    "nav",
-    `nav-${menuStyle}`,
-    loading && "nav-loading"
-  );
-
-  useEffect(() => {
-    if (menuStyle === initialStyle) {
-      return;
-    }
-
-    setLoading(true);
-    setInitialStyle(menuStyle);
-    let timer = setTimeout(() => setLoading(false), 400);
-
-    return () => {
-      clearTimeout(timer);
-    };
-    // initialStyle counts only for initial render
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [menuStyle]);
+  const { isValid: isLegalNoticeValid, url: legalNoticeUrl } = useLinkCreator({
+    path: pages.legalNotice,
+    electronFix: true,
+  });
+  const { isValid: isTermsAndConditionsValid, url: termsAndConditionsUrl } =
+    useLinkCreator({ path: pages.termsAndConditions });
+  const { isValid: isPrivacyPolicyValid, url: privacyPolicyUrl } =
+    useLinkCreator({ path: pages.privacyPolicy });
 
   return (
-    <div className={navLayoutClasses}>
+    <div className="nav">
       <ul className="button-list">
         <NavButton
           name={t("nav.learn")}
-          design={menuStyle}
           icon={<LocalLibraryIcon />}
           link="/learn"
           exact
         />
         <NavButton
           name={t("nav.newVocab")}
-          design={menuStyle}
           icon={<AddIcon />}
           link="/addVocab"
         />
         <NavButton
           name={t("nav.allVocabulary")}
-          design={menuStyle}
           icon={<LibraryBooksIcon />}
           link="/library"
         />
         <NavButton
           name={t("nav.progress")}
-          design={menuStyle}
           icon={<BarChartIcon />}
           link="/progress"
         />
-        <NavButton
-          name={t("nav.custom")}
-          design={menuStyle}
-          icon={<StyleIcon />}
-          link="/custom"
-        />
+        <NavButton name={t("nav.custom")} icon={<StyleIcon />} link="/custom" />
       </ul>
 
-      <div>
-        <NavButton
-          name={t("nav.settings")}
-          design={menuStyle}
-          icon={<SettingsIcon />}
-          link="/settings"
-        />
+      <div className="nav-legal">
+        {isLegalNoticeValid && (
+          <div className="nav-legal-wrapper">
+            <VerifiedUserIcon />
+            <a target="_blank" href={legalNoticeUrl} rel="noreferrer">
+              {t("global.legalNotice")}
+            </a>
+          </div>
+        )}
+
+        {isTermsAndConditionsValid && (
+          <div className="nav-legal-wrapper">
+            <PolicyIcon />
+            <a target="_blank" href={termsAndConditionsUrl} rel="noreferrer">
+              {t("global.termsAndConditions")}
+            </a>
+          </div>
+        )}
+
+        {isPrivacyPolicyValid && (
+          <div className="nav-legal-wrapper">
+            <SecurityIcon />
+            <a target="_blank" href={privacyPolicyUrl} rel="noreferrer">
+              {t("global.privacyPolicy")}
+            </a>
+          </div>
+        )}
       </div>
     </div>
   );