+++ src/i18n/I18nProvider.js
@@ -5,8 +5,9 @@ import { useSelector } from "react-redux";
 
 import de from "./locales/de/default.json";
 import en from "./locales/en/default.json";
-import pl from "./locales/pl/default.json";
-import ru from "./locales/ru/default.json";
+
+// import pl from "./locales/pl/default.json";
+// import ru from "./locales/ru/default.json";
 
 i18n.use(initReactI18next).init({
   debug: false,
@@ -19,8 +20,8 @@ i18n.use(initReactI18next).init({
   resources: {
     de: { default: de },
     en: { default: en },
-    pl: { default: pl },
-    ru: { default: ru },
+    // pl: { default: pl },
+    // ru: { default: ru },
   },
 });
 
@@ -39,6 +40,6 @@ export default I18nProvider;
 export const languages = [
   { code: "en", name: "English" },
   { code: "de", name: "Deutsch (German)" },
-  { code: "pl", name: "Polski (Polish)" },
-  { code: "ru", name: "Pусский (Russian)" },
+  // { code: "pl", name: "Polski (Polish)" },
+  // { code: "ru", name: "Pусский (Russian)" },
 ];