+++ src/index.js
@@ -1,3 +1,5 @@
+import "./config.js";
+
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
@@ -7,6 +9,10 @@ import App from "./App.jsx";
 import I18nProvider from "./i18n/I18nProvider.js";
 import store from "./redux/Store/index.js";
 import reportWebVitals from "./reportWebVitals.js";
+
+import "@fontsource/roboto/300.css";
+import "@fontsource/roboto/400.css";
+import "@fontsource/roboto/700.css";
 
 ReactDOM.render(
   <React.StrictMode>