+++ src/modules/clipboard.js
@@ -1,0 +1,17 @@
+let copyToClip = () => Promise.reject();
+
+if (window.VOCASCAN_CONFIG.ENV === "electron") {
+  copyToClip = ({ text }) => {
+    return window.electron.invoke("copy-to-clip", { text });
+  };
+}
+
+if (window.VOCASCAN_CONFIG.ENV === "web") {
+  if (navigator.clipboard) {
+    copyToClip = ({ text }) => {
+      return navigator.clipboard.writeText(text);
+    };
+  }
+}
+
+export { copyToClip };