+++ src/screens/Settings/Settings.jsx
@@ -1,15 +1,12 @@
 import React, { useCallback } from "react";
 import { useTranslation } from "react-i18next";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 
 import Button from "../../Components/Button/Button.jsx";
-import Switch from "../../Components/Form/Switch/Switch.jsx";
 import LanguageSelector from "../../Components/LanguageSelector/LanguageSelector.jsx";
 
+import VersionTable from "../../Components/VersionTable/VersionTable.js";
 import { openGuide } from "../../redux/Actions/login.js";
-import { setMenuStyle } from "../../redux/Actions/setting.js";
-
-import { version as desktopVersion } from "../../../package.json";
 
 import "./Settings.scss";
 
@@ -17,14 +14,6 @@ const Settings = () => {
   const { t } = useTranslation();
 
   const dispatch = useDispatch();
-  const menuStyle = useSelector((state) => state.setting.menuStyle);
-  const serverInfo = useSelector((state) => state.login.serverInfo);
-
-  const onChangeMenu = useCallback(() => {
-    dispatch(
-      setMenuStyle({ menuStyle: menuStyle === "default" ? "fancy" : "default" })
-    );
-  }, [dispatch, menuStyle]);
 
   const reopenGuide = useCallback(() => {
     dispatch(openGuide());
@@ -33,14 +22,6 @@ const Settings = () => {
   return (
     <div className="settings-wrapper">
       <h1 className="heading">{t("screens.settings.title")}</h1>
-      <Switch
-        switcher
-        label={t("screens.settings.menu.label")}
-        optionLeft={t("screens.settings.menu.optionLeft")}
-        optionRight={t("screens.settings.menu.optionRight")}
-        onChange={onChangeMenu}
-        checked={menuStyle !== "default"}
-      />
 
       <LanguageSelector />
 
@@ -51,15 +32,9 @@ const Settings = () => {
         </Button>
       </div>
 
-      <h3>
-        {t("screens.settings.desktopVersion")} v{desktopVersion}
-      </h3>
-      {serverInfo && (
-        <h3>
-          {t("screens.settings.serverVersion")} v{serverInfo.version}
-          {serverInfo.commitRef ? ` (${serverInfo.commitRef})` : ""}
-        </h3>
-      )}
+      <h2>{t("screens.settings.versions")}</h2>
+
+      <VersionTable />
     </div>
   );
 };