+++ src/utils/index.js
@@ -1,8 +1,3 @@
-/**
- *See: https://stackoverflow.com/questions/34828722/how-can-i-make-webpack-skip-a-require/34830466
- */
-export const nodeRequire = eval("require"); // eslint-disable-line no-eval
-
 /* Scale a value from one range to another
  * Example of use:
  *
@@ -48,3 +43,16 @@ export const findLanguageByCode = (language, languages) =>
  * @returns {Promise} Promise object
  */
 export const delay = (ms) => new Promise((res) => setTimeout(res, ms));
+
+/**
+ * Calculate day difference between two dates
+ * See: https://www.geeksforgeeks.org/how-to-calculate-the-number-of-days-between-two-dates-in-javascript/
+ * @param {Date} date1 first date
+ * @param {Date} date2 second date
+ * @returns {Number}
+ */
+export const dayDateDiff = (date1, date2) => {
+  const timeDiff = date2.getTime() - date1.getTime();
+
+  return Math.floor(timeDiff / (1000 * 60 * 60 * 24));
+};